requirements.txt
asgiref==3.7.2
Brotli==1.0.9
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==5.0.1
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.2.0
gunicorn==21.2.0
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
